Fix relocatable packages that own prefix in fsm
When processing the prefix directory itself, we currently fall back to
path-based operation on the real root ("/").  This results in a failure
in fchownat(2) if we're installing a package into a relocated prefix and
don't have write permissions for the real root (i.e. when running under
a regular user).

This is hardly desired behavior, and happens because we pass a pathname
argument to the *at() system calls that is absolute ("/"), which causes
them to ignore the dirfd argument, as per API specs.

Fix by passing the relative "." instead.

With that in place, we must also adjust the removal code as unlinkat(2),
when called with the AT_REMOVEDIR flag, does the equivalent of rmdir(2),
and that fails on the "." pathname.  Instead, we need to pass the file
descriptor of the prefix's parent directory and the prefix's name, for
which we can use the newly added members in the diriter struct.

However, don't ever attempt to unlink the real root if the package owns
it but is not relocated.  This would typically only produce the spurious
EBUSY warning but it's better avoided altogether.

Listing the sole prefix in a %files section is a handy way to include
the entirety of a relocatable package that's meant to be installed by
regular users (as seen in TBD) and used to work before the fsm rework
addressing the symlink CVEs, so this is technically a regression fix.

Fixes: TBD

+++ lib/fsm.c
@@ -85,7 +85,7 @@ static int fsmClose(int *wfdp);
 static char * fsmFsPath(rpmfi fi, const char * suffix)
 {
     const char *bn = rpmfiBN(fi);
-    return rstrscat(NULL, *bn ? bn : "/", suffix ? suffix : "", NULL);
+    return rstrscat(NULL, *bn ? bn : ".", suffix ? suffix : "", NULL);
 }
 
 static int fsmLink(int odirfd, const char *opath, int dirfd, const char *path)
@@ -280,9 +280,25 @@ static int fsmStat(int dirfd, const char *path, int dolstat, struct stat *sb)
     return rc;
 }
 
-static int fsmRmdir(int dirfd, const char *path)
+static int fsmRmdir(struct diriter_s di, const char *path)
 {
-    int rc = unlinkat(dirfd, path, AT_REMOVEDIR);
+    int rc = 0;
+    int dirfd = di.dirfd;
+
+    /* Handle removal of prefix dir itself */
+    if (rstreq(path, ".")) {
+	if (rstreq(di.dirfn, "/"))
+	    /* Don't even try */
+	    return rc;
+	/* Avoid EINVAL in unlinkat(2) */
+	if (di.parentdir) {
+	    dirfd = di.parentdir;
+	    path = di.dirfn;
+	}
+    }
+
+    rc = unlinkat(dirfd, path, AT_REMOVEDIR);
+
     if (_fsm_debug)
 	rpmlog(RPMLOG_DEBUG, " %8s (%d %s) %s\n", __func__,
 	       dirfd, path, (rc < 0 ? strerror(errno) : ""));
@@ -559,9 +575,9 @@ static int fsmRename(int odirfd, const char *opath, int dirfd, const char *path)
     return rc;
 }
 
-static int fsmRemove(int dirfd, const char *path, mode_t mode)
+static int fsmRemove(struct diriter_s di, const char *path, mode_t mode)
 {
-    return S_ISDIR(mode) ? fsmRmdir(dirfd, path) : fsmUnlink(dirfd, path);
+    return S_ISDIR(mode) ? fsmRmdir(di, path) : fsmUnlink(di.dirfd, path);
 }
 
 static int fsmChown(int fd, int dirfd, const char *path, mode_t mode, uid_t uid, gid_t gid)
@@ -1111,7 +1127,7 @@ setmeta:
 		continue;
 
 	    if (fp->stage > FILE_NONE && !fp->skip) {
-		(void) fsmRemove(di.dirfd, fp->fpath, fp->sb.st_mode);
+		(void) fsmRemove(di, fp->fpath, fp->sb.st_mode);
 	    }
 	}
     }
@@ -1169,7 +1185,7 @@ int rpmPackageFilesRemove(rpmts ts, rpmte te, rpmfiles files,
         if (fp->action == FA_ERASE) {
 	    int missingok = (rpmfiFFlags(fi) & (RPMFILE_MISSINGOK | RPMFILE_GHOST));
 
-	    rc = fsmRemove(di.dirfd, fp->fpath, fp->sb.st_mode);
+	    rc = fsmRemove(di, fp->fpath, fp->sb.st_mode);
 
 	    /*
 	     * Missing %ghost or %missingok entries are not errors.
